requirements.txt
playwright==1.45.1
requests==2.32.3
pydantic==2.8.2
sqlmodel==0.0.19
python-dotenv==1.0.1
tenacity==8.5.0
beautifulsoup4==4.12.3
lxml==5.3.0
aiofiles==24.1.0
psutil==6.0.0
Flask==3.0.3

# Optional: For ChatGPT API integration
# Uncomment to enable AI-enhanced scoring
# openai>=1.40.0
# tiktoken>=0.7.0